pyproject.toml
[tool.poetry]
name = "level2-3-cv-finalproject-cv-06"
version = "0.1.0"
description = ""
authors = ["백광현_T6074 <toiquen419@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
torch = {url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp311-cp311-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp311-cp311-linux_x86_64.whl"}
fastapi = {extras = ["all"], version = "^0.105.0"}
uvicorn = "^0.27.1"
loguru = "^0.7.2"
jinja2 = "^3.1.3"
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
sqlalchemy = "2.0.23"
pymysql = "^1.1.0"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
python-jose = "^3.3.0"
boto3 = "^1.34.53"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
black = "^24.2.0"
autoflake = "^2.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.flake8]
ignore = "E203, E501, W503"
max-line-length = 88
exclude = ".git,__pycache__,docs/,old/,build/,dist/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
